requirements.txt
ccxt
pandas
pandas-ta
plotly
python-dotenv
web3
